requirements.txt
# Backend — Python 3.12+
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
celery[redis]>=5.4.0

# AI Engine — LangGraph, RAG, LLM APIs
langgraph>=0.2.0
langchain-core>=0.3.0
llama-index>=0.11.0
openai>=1.0.0
anthropic>=0.39.0

# Graph DB
neo4j>=5.0.0

# Vector DB (choose one or both)
pymilvus>=2.4.0
pinecone-client>=5.0.0

# Relational DB
asyncpg>=0.29.0
sqlalchemy[asyncio]>=2.0.0

# Cache
redis>=5.0.0

# HTTP & config
httpx>=0.27.0
python-dotenv>=1.0.0
langsmith>=0.3.0

# Optional: monitoring / LLM call logging
# prometheus-client>=0.20.0
